Refactors and mobile review
