+++ src/components/Grants/GrantCard/CliffProgress.css
