+++ src/components/Grants/GrantCard/CliffProgress.tsx
@@ -3,8 +3,8 @@ import React from 'react'
 import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
 import Time from 'decentraland-gatsby/dist/utils/date/Time'
 
-import { CLIFF_PERIOD_IN_DAYS } from '../../entities/Proposal/utils'
-import '../Modal/VotingPowerDelegationDetail/VotingPowerDistribution.css'
+import { CLIFF_PERIOD_IN_DAYS } from '../../../entities/Proposal/utils'
+import '../../Modal/VotingPowerDelegationDetail/VotingPowerDistribution.css'
 
 import './CliffProgress.css'
