+++ src/components/Grants/GrantCard/GrantCardHeadline.tsx
@@ -3,8 +3,8 @@ import React from 'react'
 import TokenList from 'decentraland-gatsby/dist/utils/dom/TokenList'
 import { Header } from 'decentraland-ui/dist/components/Header/Header'
 
-import { GrantWithUpdateAttributes } from '../../entities/Proposal/types'
-import Username from '../User/Username'
+import { GrantWithUpdateAttributes } from '../../../entities/Proposal/types'
+import Username from '../../User/Username'
 
 import './GrantCardHeadline.css'
