+++ src/components/Grants/GrantCard/VestingProgress.tsx
@@ -3,8 +3,8 @@ import React from 'react'
 import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
 import Time from 'decentraland-gatsby/dist/utils/date/Time'
 
-import { GrantAttributes } from '../../entities/Proposal/types'
-import '../Modal/VotingPowerDelegationDetail/VotingPowerDistribution.css'
+import { GrantAttributes } from '../../../entities/Proposal/types'
+import '../../Modal/VotingPowerDelegationDetail/VotingPowerDistribution.css'
 
 import PercentageLabel from './PercentageLabel'
 import './VestingProgress.css'
