+++ src/components/Home/HomeSectionHeader.css
@@ -1,0 +1,31 @@
+.HomeSectionHeader__Title {
+  font-style: normal;
+  font-weight: 600;
+  font-size: 17px;
+  line-height: 28px;
+  display: flex;
+  align-items: flex-end;
+  letter-spacing: 0.4px;
+  margin-bottom: 1px;
+}
+
+.HomeSectionHeader__Description {
+  font-style: normal;
+  font-weight: 400;
+  font-size: 15px;
+  line-height: 24px;
+}
+
+@media (min-width: 768px) {
+  .HomeSectionHeader__Title {
+    margin-bottom: 4px;
+    font-size: 20px;
+  }
+
+  .HomeSectionHeader__Description {
+    width: 60%;
+    margin-bottom: 35px;
+    font-size: 17px;
+    line-height: 28px;
+  }
+}
